FS-5062 - Initial restructuring of test suite
