+++ tests/test_data/asset-information.json
