+++ tests/test_data/required-all-components.json
