+++ tests/unit/app/blueprints/fund/test_routes.py
@@ -6,7 +6,7 @@ from flask import g
 
 from app.db.models import Fund, FundingType
 from app.db.queries.fund import get_fund_by_id
-from tests.unit.helpers import submit_form
+from tests.helpers import submit_form
 
 
 @pytest.mark.usefixtures("set_auth_cookie", "patch_validate_token_rs256_internal_user")
